keycloak-lab/demo_api/requirements.txt
flask
pyjwt[crypto]
requests
